@@ -2,8 +2,8 @@ import { Buffer } from 'buffer';
 import createBuffer from './create_buffer';
 import defineCrc from './define_crc';
 
-const crc1 = defineCrc('crc1', function(buf, previous) {
-  if (!Buffer.isBuffer(buf)) buf = createBuffer(buf);
+const crc1 = defineCrc('crc1', (value, previous) => {
+  const buf = Buffer.isBuffer(value) ? value : createBuffer(value);
 
   let crc = ~~previous;
   let accum = 0;
