@@ -36,13 +36,13 @@ let TABLE = [
   0x8801, 0x48c0, 0x4980, 0x8941, 0x4b00, 0x8bc1, 0x8a81, 0x4a40,
   0x4e00, 0x8ec1, 0x8f81, 0x4f40, 0x8d01, 0x4dc0, 0x4c80, 0x8c41,
   0x4400, 0x84c1, 0x8581, 0x4540, 0x8701, 0x47c0, 0x4680, 0x8641,
-  0x8201, 0x42c0, 0x4380, 0x8341, 0x4100, 0x81c1, 0x8081, 0x4040
-]
+  0x8201, 0x42c0, 0x4380, 0x8341, 0x4100, 0x81c1, 0x8081, 0x4040,
+];
 
 if (typeof Int32Array !== 'undefined') TABLE = new Int32Array(TABLE);
 
-const crc16 = defineCrc('crc-16', function(buf, previous) {
-  if (!Buffer.isBuffer(buf)) buf = createBuffer(buf);
+const crc16 = defineCrc('crc-16', (value, previous) => {
+  const buf = Buffer.isBuffer(value) ? value : createBuffer(value);
 
   let crc = ~~previous;
 
