@@ -36,13 +36,13 @@ let TABLE = [
   0xfd2e, 0xed0f, 0xdd6c, 0xcd4d, 0xbdaa, 0xad8b, 0x9de8, 0x8dc9,
   0x7c26, 0x6c07, 0x5c64, 0x4c45, 0x3ca2, 0x2c83, 0x1ce0, 0x0cc1,
   0xef1f, 0xff3e, 0xcf5d, 0xdf7c, 0xaf9b, 0xbfba, 0x8fd9, 0x9ff8,
-  0x6e17, 0x7e36, 0x4e55, 0x5e74, 0x2e93, 0x3eb2, 0x0ed1, 0x1ef0
+  0x6e17, 0x7e36, 0x4e55, 0x5e74, 0x2e93, 0x3eb2, 0x0ed1, 0x1ef0,
 ];
 
 if (typeof Int32Array !== 'undefined') TABLE = new Int32Array(TABLE);
 
-const crc16ccitt = defineCrc('ccitt', function(buf, previous) {
-  if (!Buffer.isBuffer(buf)) buf = createBuffer(buf);
+const crc16ccitt = defineCrc('ccitt', (value, previous) => {
+  const buf = Buffer.isBuffer(value) ? value : createBuffer(value);
 
   let crc = typeof previous !== 'undefined' ? ~~previous : 0xffff;
 
