@@ -1,5 +1,5 @@
-export default function(model, calc) {
-  const fn = (buf, previous) => calc(buf, previous) >>> 0;
+export default function defineCrc(model, calc) {
+  const fn = (value, previous) => calc(value, previous) >>> 0;
   fn.signed = calc;
   fn.unsigned = fn;
   fn.model = model;
