@@ -2,7 +2,7 @@ import crcSuiteFor from './test_helpers';
 import crc16 from '../lib/es6/crc16';
 import createBuffer from '../lib/es6/create_buffer';
 
-describe('CRC16', function() {
+describe('CRC16', () => {
   crcSuiteFor({ crc: crc16 });
 
   // https://github.com/alexgorbatchev/node-crc/issues/29
