@@ -1,6 +1,6 @@
 import crcSuiteFor from './test_helpers';
 import crcjam from '../lib/es6/crcjam';
 
-describe('CRCJAM', function() {
+describe('CRCJAM', () => {
   crcSuiteFor({ crc: crcjam });
 });
